// src/interfaces/CarInterface.tsx
export interface ResponseCar {
  ok: boolean
  msg: string
  data: Data
}

export interface Data {
  driver_id: number
  plate: string
  color: string
  model: string
  hcl: number
  id: number
  createdAt: string
  updatedAt: string
  photo_car: string
  photo_car_side:                string;
  photo_car_back:                string;
  photo_registration_front: string
}

// Generated by https://quicktype.io

export interface ResponseMyCars {
  ok:   boolean;
  msg:  string;
  data: DatumMyCars[];
}

export interface DatumMyCars {
  id:                       number;
  driver_id:                number;
  plate:                    string;
  photo_car:                string;
  photo_car_side:                string;
  photo_car_back:                string;
  photo_registration_front: string;
  photo_registration_back:  string;
  color:                    string;
  model:                    string;
  hcl:                      string;
  is_active:                boolean;
  extras:[{
    car_id:number;
    extra_service_id:number;
    id:number;
  }]
}


// Generated by https://quicktype.io

export interface ResponseSelectMyCar {
  ok:   boolean;
  msg:  string;
  data: Data;
}

export interface Data {
  id:                       number;
  driver_id:                number;
  plate:                    string;
  photo_car:                string;
  photo_car_side:                string;
  photo_car_back:                string;
  photo_registration_front: string;
  photo_registration_back:  string;
  color:                    string;
  model:                    string;
  hcl:                      any;
  is_active:                boolean;
  selected:                 boolean;
}